limit changDiscountPage input ,hashtag tidy in a folder
