+++ frontend/src/lib/components/PUA/hashtag/hashtagAdd.svelte
@@ -1,0 +1,6 @@
+<button
+	on:click
+	class="flex h-7 w-7 items-center justify-center rounded-full bg-PUA-dark-red text-center text-xl font-bold text-white"
+>
+	+
+</button>
